Add an error boundary to the embed page
This change seems larger than it is mostly because I had to move
some stuff around to get rid of a dependency cycle in our module
structure which we unfortunately have lots of.

This was causing load order inconsistencies and made the app crash.
