+++ frontend/src/routes/Upload.tsx
@@ -13,14 +13,14 @@ import { UPLOAD_PATH } from "./paths";
 import { makeRoute } from "../rauta";
 import { UploadJwtQuery } from "./__generated__/UploadJwtQuery.graphql";
 import { assertNever, bug, ErrorDisplay, errorDisplayInfo, unreachable } from "../util/err";
-import { currentRef, useNavBlocker } from "../util";
+import { useNavBlocker } from "./util";
 import CONFIG from "../config";
 import { Button } from "../ui/Button";
 import { boxError, ErrorBox } from "../ui/error";
 import { Form } from "../ui/Form";
 import { Input, TextArea } from "../ui/Input";
 import { User, useUser } from "../User";
-import { useRefState } from "../util";
+import { currentRef, useRefState } from "../util";
 import { Card } from "../ui/Card";
 import { InputContainer, TitleLabel } from "../ui/metadata";
 import { PageTitle } from "../layout/header/ui";
